refact: Move constants to new file and use MRepo
Moves constants from component_utils to constants.py
so that ModulesRepo can be used from inside of component_utils,
thereby avoiding a circular import
