+++ nf_core/components/components_utils.py
@@ -1,24 +1,17 @@
 import logging
 import re
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import questionary
 import requests
 import rich.prompt
 import yaml
 
-if TYPE_CHECKING:
-    from nf_core.modules.modules_repo import ModulesRepo
-
 import nf_core.utils
+from nf_core.modules.modules_repo import ModulesRepo
 
 log = logging.getLogger(__name__)
-
-# Constants for the nf-core/modules repo used throughout the module files
-NF_CORE_MODULES_NAME = "nf-core"
-NF_CORE_MODULES_REMOTE = "https://github.com/nf-core/modules.git"
-NF_CORE_MODULES_DEFAULT_BRANCH = "master"
 
 
 def get_repo_info(directory: Path, use_prompt: Optional[bool] = True) -> Tuple[Path, Optional[str], str]:
@@ -181,14 +174,14 @@ def get_components_to_install(
                 for component in components:
                     if isinstance(component, dict):
                         component_name = list(component.keys())[0].lower()
-                        git_remote = component[component_name]["git_remote"]
-                        org_path = Path(subworkflow_dir).parent.stem
+                        branch = component[component_name].get("branch")
+                        modules_repo = ModulesRepo(component[component_name]["git_remote"], branch)
                         current_comp_dict = subworkflows if component_name in subworkflows else modules
 
                         component_dict = {
-                            "org_path": org_path,
-                            "git_remote": git_remote,
-                            "branch": component[component_name].get("branch"),
+                            "org_path": modules_repo.repo_path,
+                            "git_remote": modules_repo.remote_url,
+                            "branch": branch,
                         }
 
                         current_comp_dict[component_name].update(component_dict)
