+++ nf_core/components/constants.py
@@ -1,0 +1,4 @@
+# Constants for the nf-core/modules repo used throughout the module files
+NF_CORE_MODULES_NAME = "nf-core"
+NF_CORE_MODULES_REMOTE = "https://github.com/nf-core/modules.git"
+NF_CORE_MODULES_DEFAULT_BRANCH = "master"
