+++ nf_core/synced_repo.py
@@ -9,7 +9,7 @@ from typing import Dict, Iterable, List, Optional, Union
 import git
 from git.exc import GitCommandError
 
-from nf_core.components.components_utils import (
+from nf_core.components.constants import (
     NF_CORE_MODULES_NAME,
     NF_CORE_MODULES_REMOTE,
 )
